# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SUSE AI Universal Proxy provides RESTful APIs for managing MCP (Model Context Protocol) server deployments and proxying requests to MCP servers running in Kubernetes.",
        "title": "SUSE AI Universal Proxy - Control Plane",
        "contact": {
            "name": "SUSE AI Universal Proxy Team",
            "url": "https://github.com/SUSE/suse-ai-up"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/SUSE/suse-ai-up/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8911",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/adapters/{name}/token": {
            "get": {
                "description": "Retrieves the current token for an adapter or generates a new one if none exists",
                "tags": [
                    "tokens"
                ],
                "summary": "Get adapter token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Generate new token if none exists",
                        "name": "generate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 24,
                        "description": "Token expiration time in hours",
                        "name": "expiresIn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adapters/{name}/token/refresh": {
            "post": {
                "description": "Generates a new token for an adapter, invalidating the old one",
                "tags": [
                    "tokens"
                ],
                "summary": "Refresh adapter token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 24,
                        "description": "Token expiration time in hours",
                        "name": "expiresIn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/adapters/{name}/token/validate": {
            "post": {
                "description": "Validates a token and returns its claims and validity",
                "tags": [
                    "tokens"
                ],
                "summary": "Validate token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adapter name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token to validate",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.TokenInfo": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "audience": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "issuedAt": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "tokenId": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}